Currently compiling ahead of VS integration.
Setup TTrack to use pointers rather than instances for Detector and
Tracker. This is to allow parameter specific constructors to be called
even when using a singleton TTrack.

Currently a bug occurs with the set of paths returned by the
Handler. This causes one path to have a junk name. No idea why this
occurs at the moment.